[refactor] Dto: group DTOs by api
