+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/repository/mapper/ConsumptionMapper.kt
@@ -7,7 +7,7 @@
 
 package com.rwmobi.kunigami.data.repository.mapper
 
-import com.rwmobi.kunigami.data.source.network.dto.ConsumptionDto
+import com.rwmobi.kunigami.data.source.network.dto.consumption.ConsumptionDto
 import com.rwmobi.kunigami.domain.extensions.roundToNearestEvenHundredth
 import com.rwmobi.kunigami.domain.model.consumption.Consumption
