+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/account/AccountApiResponse.kt
@@ -5,13 +5,13 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.account
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class MeterDto(
-    @SerialName("serial_number") val serialNumber: String,
-    @SerialName("registers") val registers: List<RegisterDto>? = null,
+data class AccountApiResponse(
+    @SerialName("number") val number: String,
+    @SerialName("properties") val properties: List<PropertyDto>,
 )
