+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/account/ElectricityMeterPointDto.kt
@@ -5,15 +5,17 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.account
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GasMeterPointDto(
-    @SerialName("mprn") val mprn: String,
+data class ElectricityMeterPointDto(
+    @SerialName("mpan") val mpan: String,
+    @SerialName("profile_class") val profileClass: Int,
     @SerialName("consumption_standard") val consumptionStandard: Int,
     @SerialName("meters") val meters: List<MeterDto>,
     @SerialName("agreements") val agreements: List<AgreementDto>,
+    @SerialName("is_export") val isExport: Boolean,
 )
