+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/account/GasMeterPointDto.kt
@@ -5,17 +5,15 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.account
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ElectricityMeterPointDto(
-    @SerialName("mpan") val mpan: String,
-    @SerialName("profile_class") val profileClass: Int,
+data class GasMeterPointDto(
+    @SerialName("mprn") val mprn: String,
     @SerialName("consumption_standard") val consumptionStandard: Int,
     @SerialName("meters") val meters: List<MeterDto>,
     @SerialName("agreements") val agreements: List<AgreementDto>,
-    @SerialName("is_export") val isExport: Boolean,
 )
