+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/consumption/ConsumptionApiResponse.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.consumption
 
 import kotlinx.serialization.Serializable
