+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/prices/PricesApiResponse.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.prices
 
 import kotlinx.serialization.Serializable
