+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/prices/RateDto.kt
@@ -5,15 +5,17 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.prices
 
 import kotlinx.datetime.Instant
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class AgreementDto(
-    @SerialName("tariff_code") val tariffCode: String,
+data class RateDto(
+    @SerialName("value_exc_vat") val vatExclusivePrice: Double,
+    @SerialName("value_inc_vat") val vatInclusivePrice: Double,
     @SerialName("valid_from") val validFrom: Instant,
     @SerialName("valid_to") val validTo: Instant?,
+    @SerialName("payment_method") val paymentMethod: String?,
 )
