+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/data/source/network/dto/singleproduct/ElectricityTariffDto.kt
@@ -5,8 +5,9 @@
  *
  */
 
-package com.rwmobi.kunigami.data.source.network.dto
+package com.rwmobi.kunigami.data.source.network.dto.singleproduct
 
+import com.rwmobi.kunigami.data.source.network.dto.LinkDto
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
