+++ composeApp/src/commonMain/kotlin/com/rwmobi/kunigami/ui/viewmodels/TariffsViewModel.kt
@@ -61,6 +61,9 @@ class TariffsViewModel(
         }
     }
 
+    fun getTariffDetails(productCode: String) {
+    }
+
     fun requestScrollToTop(enabled: Boolean) {
         _uiState.update { currentUiState ->
             currentUiState.copy(
