+++ composeApp/src/commonTest/kotlin/com/rwmobi/kunigami/data/source/network/samples/GetAccountSampleData.kt
@@ -7,13 +7,13 @@
 
 package com.rwmobi.kunigami.data.source.network.samples
 
-import com.rwmobi.kunigami.data.source.network.dto.AccountApiResponse
-import com.rwmobi.kunigami.data.source.network.dto.AgreementDto
-import com.rwmobi.kunigami.data.source.network.dto.ElectricityMeterPointDto
-import com.rwmobi.kunigami.data.source.network.dto.GasMeterPointDto
-import com.rwmobi.kunigami.data.source.network.dto.MeterDto
-import com.rwmobi.kunigami.data.source.network.dto.PropertyDto
-import com.rwmobi.kunigami.data.source.network.dto.RegisterDto
+import com.rwmobi.kunigami.data.source.network.dto.account.AccountApiResponse
+import com.rwmobi.kunigami.data.source.network.dto.account.AgreementDto
+import com.rwmobi.kunigami.data.source.network.dto.account.ElectricityMeterPointDto
+import com.rwmobi.kunigami.data.source.network.dto.account.GasMeterPointDto
+import com.rwmobi.kunigami.data.source.network.dto.account.MeterDto
+import com.rwmobi.kunigami.data.source.network.dto.account.PropertyDto
+import com.rwmobi.kunigami.data.source.network.dto.account.RegisterDto
 import kotlinx.datetime.Instant
 
 object GetAccountSampleData {
